FIX: parallel write in dentk-swapaxes
Individual threads create individual wirtters and write different blocks
of output file to be more efficient.

In the original implementation when the file was not created, multiple
threads tried to create it leading to the KCTException.

Empty file is now created before parallel write so that the threads work
with this file but do not attempt to create it.