dentk-calc1 adding function doublesoftplus
Softplus produces zeros for ~ -30,  softplus(softplus(0))~1,
softplus(softplus(30))~30

For this properties is convenient to use double softplus e.g. for flat
field correction for the preprocessing of the flat fields as those are
typically the numbers with which is then divided

So it is not suffucient one softplus for regularization but two softplus
functions actually do the trick of being always larger than zeros, do
not having flat single value areas and being almost the same for most of
the values.